Align kernel and user-space view of stat_t.

+++ include/common.h
@@ -22,7 +22,7 @@ typedef uint32_t mode_t;
 typedef uint16_t nlink_t;
 typedef uint16_t ino_t;
 typedef uint32_t tid_t;
-typedef int64_t blkcnt_t;  /* fs block count */
+typedef int32_t blkcnt_t;  /* fs block count */
 typedef int32_t blksize_t; /* fs optimal block size */
 
 /* Generic preprocessor macros */
